added unprotected components page
