+++ pages/index.tsx
@@ -1,41 +1,11 @@
-import CustomizeDashboard from '@/components/CustomizeDashboard';
 import LoginStatus from '@/components/LoginStatus';
-import { Notification } from '@/src/models/Notification';
 import protectedGetServerSideProps from '@/src/utils/protectRoute';
 import Head from 'next/head';
-import NotificationBubble from '../components/NotificationBubble';
-import SearchBar from '../components/SearchBar';
-import SideNav from '../components/SideNav';
+import Link from 'next/link';
 
 export const getServerSideProps = protectedGetServerSideProps;
 
 export default function Home() {
-  const notifications: Notification[] = [
-    {
-      icon: 'mingcute:trophy-fill',
-      messageHighlight: 'Wowow you just got a trophy!',
-      mainMessage: '',
-      timestamp: 25,
-      partySize: undefined,
-      maxPartySize: undefined,
-    },
-    {
-      icon: 'mdi:thumb-up',
-      messageHighlight: '@thisandthat',
-      mainMessage: 'from liked your post',
-      timestamp: 62,
-      partySize: undefined,
-      maxPartySize: undefined,
-    },
-    {
-      icon: 'icon-park-solid:game-handle',
-      messageHighlight: '@soandso',
-      mainMessage: 'joined your game lobby',
-      timestamp: 134,
-      partySize: 3,
-      maxPartySize: 4,
-    },
-  ];
   return (
     <>
       <Head>
@@ -46,12 +16,10 @@ export default function Home() {
       </Head>
       <main>
         <LoginStatus />
-        <SideNav />
-        <SearchBar />
-        <CustomizeDashboard />
-        {notifications.map((notification, index) => (
-          <NotificationBubble key={index} notification={notification} />
-        ))}
+        <h1>Hello! You are currently looking at a protected route</h1>
+        <Link href={"/components"}>
+          <button>See components page</button>
+        </Link>
       </main>
     </>
   );
